+++ src/App/Controller/ProjectController.php
@@ -2,7 +2,7 @@
 
 namespace App\Controller;
 
-use App\Entity\Repository\ProjectRepository;
+use App\Repository\ProjectRepository;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Attribute\Route;
