+++ src/App/Entity/Package.php
@@ -2,7 +2,7 @@
 
 namespace App\Entity;
 
-use App\Entity\Repository\PackageRepository;
+use App\Repository\PackageRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
