+++ src/App/Entity/Project.php
@@ -2,7 +2,7 @@
 
 namespace App\Entity;
 
-use App\Entity\Repository\ProjectRepository;
+use App\Repository\ProjectRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
