+++ src/App/Resources/config/services.yml
@@ -8,5 +8,5 @@ services:
         exclude:
             - '../../Tests/'
             - '../../DependencyInjection/'
-            - '../../Entity/*.php'
+            - '../../Entity/'
             - '../../Kernel.php'
